#560 [feature] Add Main, Home Navigation
