+++ presentation/src/main/java/daily/dayo/presentation/screen/feed/FeedNavigation.kt
@@ -1,0 +1,15 @@
+package daily.dayo.presentation.screen.feed
+
+import androidx.compose.material.ExperimentalMaterialApi
+import androidx.navigation.NavGraphBuilder
+import androidx.navigation.compose.composable
+import daily.dayo.presentation.screen.main.Screen
+
+@OptIn(ExperimentalMaterialApi::class)
+fun NavGraphBuilder.feedNavGraph(
+
+) {
+    composable(Screen.Feed.route) {
+        // FeedScreen()
+    }
+}
