+++ presentation/src/main/java/daily/dayo/presentation/screen/main/MainNavigator.kt
@@ -1,0 +1,29 @@
+package daily.dayo.presentation.screen.main
+
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.remember
+import androidx.navigation.NavDestination
+import androidx.navigation.NavHostController
+import androidx.navigation.compose.currentBackStackEntryAsState
+import androidx.navigation.compose.rememberNavController
+
+internal class MainNavigator(
+    val navController: NavHostController,
+) {
+    private val currentDestination: NavDestination?
+        @Composable get() = navController
+            .currentBackStackEntryAsState().value?.destination
+
+    @Composable
+    fun shouldShowBottomBar(): Boolean {
+        val currentRoute = currentDestination?.route ?: return false
+        return currentRoute in Screen
+    }
+}
+
+@Composable
+internal fun rememberMainNavigator(
+    navController: NavHostController = rememberNavController(),
+): MainNavigator = remember(navController) {
+    MainNavigator(navController)
+}
